Modification type of variables

+++ puissance4.c
@@ -5,7 +5,7 @@
 #include "minmax.h"
 
 
-int grid[NB_LINE][NB_COLUMN] = {{0}};
+uint8_t grid[NB_LINE][NB_COLUMN] = {{0}};
 
 
 void display_grid(void) {
@@ -32,7 +32,7 @@ void display_grid(void) {
 				continue;
 			}
 			if (grid[i][j-1] != 0) {
-				printf(" %i |", grid[i][j-1]);
+				printf(" %u |", grid[i][j-1]);
 			} else {
 				printf("   |");
 			}
@@ -50,7 +50,7 @@ void display_grid(void) {
 }
 
 
-int is_full(int column) {
+uint8_t is_full(uint8_t column) {
 	if (grid[0][column] != 0) {
 		return 1;
 	}
@@ -58,7 +58,7 @@ int is_full(int column) {
 }
 
 
-int play(int column, int player) {
+uint8_t play(uint8_t column, uint8_t player) {
 	for (int i = 1; i < NB_LINE; i++) {
 		if (grid[i][column] != 0) {
 			grid[i-1][column] = player;
@@ -80,12 +80,12 @@ int play(int column, int player) {
 }
 
 
-int win_horizontal(int line, int player) {
-	int win = 0;
-	int count = 1;
-	int symbol = player;
-	int previous = grid[line][0];
-	int current;
+uint8_t win_horizontal(uint8_t line, uint8_t player) {
+	uint8_t win = 0;
+	uint8_t count = 1;
+	uint8_t symbol = player;
+	uint8_t previous = grid[line][0];
+	uint8_t current;
 
 	// Check on the line if 4 symbols aligned
 	for (int j = 1; j < NB_COLUMN; j++) {
@@ -107,12 +107,12 @@ int win_horizontal(int line, int player) {
 }
 
 
-int win_vertical(int column, int player) {
-	int win = 0;
-	int count = 1;
-	int symbol = player;
-	int previous = grid[0][column];
-	int current;
+uint8_t win_vertical(uint8_t column, uint8_t player) {
+	uint8_t win = 0;
+	uint8_t count = 1;
+	uint8_t symbol = player;
+	uint8_t previous = grid[0][column];
+	uint8_t current;
 
 	// Check on the column if 4 symbols aligned
 	for (int i = 1; i < NB_LINE; i++) {
@@ -134,16 +134,16 @@ int win_vertical(int column, int player) {
 }
 
 
-int win_diagonal(int line, int column, int player) {
-	int win = 0;
-	int count = 1;
-	int symbol = player;
-	int current;
-	int previous;
+uint8_t win_diagonal(uint8_t line, uint8_t column, uint8_t player) {
+	uint8_t win = 0;
+	uint8_t count = 1;
+	uint8_t symbol = player;
+	uint8_t current;
+	uint8_t previous;
 
 	// Right diagonal
-	int i = line;
-	int j = column;
+	uint8_t i = line;
+	uint8_t j = column;
 	// Find the lowest extremity of the diagonal
 	while (i < NB_LINE && j > 0) {
 		i++;
@@ -204,7 +204,7 @@ int win_diagonal(int line, int column, int player) {
 }
 
 
-int check_win(int line, int column, int player) {
+uint8_t check_win(uint8_t line, uint8_t column, uint8_t player) {
 	if (win_horizontal(line, player) || win_vertical(column, player) || win_diagonal(line, column, player)) {
 		return 1;
 	}
@@ -213,11 +213,11 @@ int check_win(int line, int column, int player) {
 
 
 int main(void) {
-	int player1_column;
+	uint8_t player1_column;
 #if 0
-	int player2_column;
+	uint8_t player2_column;
 #endif
-	int win;
+	uint8_t win;
 	
 	display_grid();
 
@@ -229,18 +229,18 @@ int main(void) {
 
 		// PLAYER 1
 		printf("Player 1 : ");
-		scanf("%d", &player1_column);
+		scanf("%hhu", &player1_column);
 		
 		// If the number of the colonne is not good
 		while (player1_column < 1 || player1_column > 7) {
 			printf("Player 1 : Error enter again the column = ");
-			scanf("%d", &player1_column);
+			scanf("%hhu", &player1_column);
 		}
 
 		// If the column is full
 		while (is_full(player1_column-1)) {
 			printf("Player 1 : Error column full, enter an other column = ");
-			scanf("%d", &player1_column);
+			scanf("%hhu", &player1_column);
 		}
 
 		// Now we can play
